+++ lib/bloom_rates.rb
@@ -17,11 +17,11 @@ module BloomRates
 
   def self.setup(channel = DEFAULT_CHANNEL)
     client = MessageBus::Client.new(BloomRates.configuration.bloom_trade_url)
-    last_id = BloomRates::MessageBusLastIdSetter.()
+    current_last_id = BloomRates::MessageBusLastIdSetter.()
 
-    client.subscribe(channel, last_id) do |payload, last_id|
+    client.subscribe(channel, current_last_id) do |payload, last_id|
       BloomRates::ExchangeRates::Sync.(payload)
-      BloomRates::MessageBusLastId.create!(last_id: last_id)
+      BloomRates::MessageBusLastId.create_or_update(last_id: last_id)
     end
 
     client.start