+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Create MessageBusLastId.create_or_update
+- Only store 1 record of `MessageBusLastId`
 
 ## [0.5.3] - 2018-07-12
 - Change client test to use `webmock`